adding transaction Api endpoint

+++ src/main/java/com/ylizma/bankmanagement/domain/TransferDetails.java
@@ -5,6 +5,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @AllArgsConstructor
 @NoArgsConstructor
 @Getter
@@ -15,5 +17,5 @@ public class TransferDetails {
 	
 	private Long toAccountNumber;
 	
-	private Double transferAmount;
+	private BigDecimal transferAmount;
 }
